server/requirements.txt
fastapi
sqlalchemy
passlib
bcrypt
pyjwt
python-dotenv